Restructure project
Created new packages an rearranged the classes in a more meaningful
way. Most of the implementation is now in processor package which
is devided into few other packages. Implementation of general
default classes are in defaults package. Language specific
implementations are in languages package and framework specific
implementations are in frameworks package.