Add support for multi-annotation processing
Added support for processing more than one annotation in a single
processor. For that feature the code and tests had to be refactored
quite heavily.

Added new OverridePatter annotation for overriding JSR-303 regexp
patter in the target model. This will add more flexibility for the
pattern annotation since there may be some cases where the converter
doesn't work as expected. Now the user can override the generated
pattern with a custom pattern. This new feature required the multi-
annotation processing.

Added tests for new classes and BaseValidationAnnotationProcessor
and DefaultAnnotationProcessorProvider.