Fix issues with non-base validation processors
Fixed issues that didn't allow non-base validation processors to be
run. This was related to using default version of
BackboneFieldAnnotationWriter instead of one given with constructor
arguments. The bug affected mainly the tests but it could have affected
production code (if only someone would use this in production someday).
Fix also enables the user to add custom validation processors
on-the-fly.