Add missing validators and writers
Added missing validators and correspnding writers. Changed validator
construction to use the default constructor and forcing the user to
implement init method. The main reason for this was to make the default
constructor to init the type automatically and making sure that the
user will create a right kind of init method. The old method that
relied on constructor with parameter made IDE to generate from kind
of constructor. The new abstract init method will be always right type.

Optimized imports in all java files