Reorder class/package structure for some classes
Moved default value processing from ModelFactory to ModelProcessing.
This will cut ModelFactory dependencies to specific language or
framework making it much more generic.

The default value processing doesn't yet work as it previsouly worked
because there is not way to give the default valut object to it.

Removed HTML5 processing classes which are replaced with Backbone
Forms support.

Move files to better packages