Add initial support for JSON printing for BB
Added initial support for outputting JSON instead of Backbones Model
extended JavaScript classes. This new mode can be used when giving the
validation data dunamically to a JS app. The models are written by hand
for the app, but the validation data is fetched when the app starts and
injected to the client side models.