Clean BackboneModelProcessor creation
Add builder for BackboneModelProcessor

Change BackboneModelProcessor to use a list of ModelProcessors instead of
predefined ModelProcessors. This makes configurating the
BackboneModelProcessor much easier.

Changed how writers are set for ModelProcessors. The old way required the
same writer to be set for multiple different objects. The new system
allows the writers to propagate to other classes when it is added to
BackboneModelProcessor.